+++ lib/DAV/Auth/Backend/Apache.php
@@ -86,7 +86,7 @@ class Apache implements BackendInterface {
      * existing one.
      *
      * @param RequestInterface $request
-     * @param ResponseInterface $request
+     * @param ResponseInterface $response
      * @return void
      */
     function challenge(RequestInterface $request, ResponseInterface $response) {