+++ lib/DAV/Xml/Response/MultiStatus.php
@@ -39,6 +39,7 @@ class MultiStatus implements Element {
      * Constructor
      *
      * @param \Sabre\DAV\Xml\Element\Response[] $responses
+     * @param string $syncToken
      */
     function __construct(array $responses, $syncToken = null) {